Added endpoint /weather/{date}, which gives max,min and avg temperature and condition for specific date

+++ src/main/java/com/example/first_spring_ib/Day.java
@@ -9,7 +9,7 @@ import lombok.Data;
 public class Day {
 
     private float maxtemp_c;
-    private double mintemp_c;
-    private double avgtemp_c;
+    private float mintemp_c;
+    private float avgtemp_c;
     private Condition condition;
 }
